feat: add finalize round  instruction

+++ programs/diamond_arena/src/constants.rs
@@ -8,8 +8,8 @@ pub const DEFAULT_LIVES: u8 = 3;
 pub const MIN_PLAYERS: u8 = 2;
 pub const MAX_PLAYERS: u8 = 5;
 
-pub const PROTOCOL_FEE_BPS: u16 = 100; // 1%
-pub const MAX_NUMBER: u16 = 100;
+pub const PROTOCOL_FEE_BPS: u8 = 100; // 1%
+pub const MAX_NUMBER: u8 = 100;
 
 pub const DISCRIMINATOR: usize = 8;
